translation, nucnce, and flow fixes
The "Best wishes!" translation is to maintain consistency with other
places where Ooishi is saying this. I think it makes sense to avoid the
literal "Happy New Year" for this line.

When I asked on the steam forum, a Japanese player was kind enough to
explain that Ryukishi07 used to use "Happy New Year" as a greeting at
his/her old workplace, and so added that quirk to Ooishi. I think it's a
fun thing, but it just doesn't make sense to me for it to be in the
English version (impossible to find this out in English afaik)