hmmm
I just noticed for the first time (having to write it out myself) that
Rena's name is spelled out weirdly.

The type of romanization employed seems to be keeping long vowels (Like
Ooishi as opposed to Oishi) but Ryuuguu is half-half - despite being two
long vowels only the first one is kept long in romanization.