leaving comments in the file
I wanted to leave an explanation for why the translation doesn't match
the Japanese here.

a pretty literal translation of Rika's line here would be "... If you
took out any one of us in remaining lines to record, Keiichi would be
able to join the party"
This just doesn't make sense to me, and didn't really get cleared up by
asking Japanese players.
So I'm keeping the Sui text in Japanese so that it isn't a weird
mismatch with the audio for Japanese players, but keeping the
translation of the original game text, because it works better IMO.